+++ src/L2Data2Code/L2Data2Code.BaseGenerator/Entities/EntityColumn.cs
@@ -29,6 +29,7 @@ namespace L2Data2Code.BaseGenerator.Entities
         public string DbToField { get; set; }
         public string DefaultValue { get; set; }
         public bool HasDefaultValue => !string.IsNullOrWhiteSpace(DefaultValue);
+        public bool HasRelation { get; set; }
 
         public string GetDefaultValue()
         {