+++ src/L2Data2Code/L2Data2Code.Avalonia/changes.md
@@ -1,3 +1,8 @@
+### 3.3.8
+
+- Added property `HasRelation` to `Entity` and `NotRelatedColumns` for replacement
+- Added `FieldDescriptor` and `FirstPK` to `Entity`
+
 ### 3.3.7
 
 - Update some packages